CloudEvents support (#252)
CloudEvents is an essential choice for Liiklus since it is the standard that we were missing when we created Liiklus.

It introduces a clear notion of id, type and other fields that can be helpful for applying schema, RBAC, metrics and many other features

At this stage, CloudEvents v1 is used as internal representation, and `LiiklusEvent` at the API level since there is no Protobuf format for CloudEvents yet:
https://github.com/cloudevents/spec/pull/504
`LiiklusEvent` follows the v1 of the spec. Once Protobuf definitions are available, we will deprecate `LiiklusEvent` and start using CloudEvent message type. oneof and format are there for a smooth migration.

The implementation requires you to provide a plugin to "upcast" old events (use `RecordPreProcessor`/`RecordPostProcessor`). See `CloudEventsTest` for an example of it.