CVE-2019-15052 and CVE-2019-16370 fix with gradle update to 6.3 (#288)
* CVE-2019-15052 and CVE-2019-16370 fix with gradle update

The PGP signing plugin in Gradle before 6.0 relies on the SHA-1 algorithm, which might allow an attacker to replace an artifact with a different one that has the same SHA-1 message digest, a related issue to CVE-2005-4900.

The HTTP client in Gradle before 5.6 sends authentication credentials originally destined for the configured host. If that host returns a 30x redirect, Gradle also sends those credentials to all subsequent hosts that the request redirects to. This is similar to CVE-2018-1000007.

* gradle 6.3